Simplification & fixed formatting. #258

+++ mz_zip.h
@@ -47,7 +47,7 @@ typedef struct mz_zip_file_s
     uint64_t disk_offset;               // relative offset of local header
     uint16_t internal_fa;               // internal file attributes
     uint32_t external_fa;               // external file attributes
-    uint32_t zip64;                     // zip64 extension mode
+    uint16_t zip64;                     // zip64 extension mode
 
     char     *filename;                 // filename string
     uint8_t  *extrafield;               // extrafield data
